dev Refactor command handling: replace UndoableCommand with IrreversibleCommand and update related tests
